WIP: Move initial implementation into V1, fiddling with PATCH
The PATCH implementation is currently an experiement in dynamically
interacting with models, inspired by Selcuk's Team Data API ;)

Moving the basic CRUD implementation into V1 is so that the naive
approach can be separated there, while a more clever approach can
be implemented alongside as "V2". This is good exercise for evolving
APIs, as well.