Update postman collection. Add testing docco and scripts
If nothing else, this is an exercise in using Postman to define
API tests and then run those tests from a command line using
Newman.

Making a "it has some usable data in it" commit of _db.sqlite3
to the repo so that my Pycharm run configuration to run the Newman
tests can be updated to reset the db file each time without loosing
the helpful existing data.